CPP Check and add docs

+++ src/engine/helpers/stringFns.h
@@ -16,7 +16,7 @@ extern float toFloat(std::string str);
 
 template <typename... Args>
 std::string format(const std::string& format, Args... args) {
-  size_t size = snprintf(nullptr, 0, format.c_str(), args...) + 1;
+  int size = snprintf(nullptr, 0, format.c_str(), args...) + 1;
   if (size <= 0) {
     throw std::runtime_error("Error during formatting.");
   }
